requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
pytest>=7.4.0
ddtrace>=2.0.0
datadog>=0.47.0
psutil>=5.9.0
